requirements.txt
kafka==1.3.5
pyOpenSSL==17.5.0